refactor + update docs
