+++ docs/installation.rst
@@ -1,0 +1,10 @@
+Installation
+============
+
+Use the package manager `pip <https://pip.pypa.io/en/stable/>`_ to install pqr.
+
+.. code:: bash
+
+    pip install pqr
+
+There are only 2 dependencies: numpy and pandas.
