+++ pqr/allocation.py
